extracted graph building to external function
